# backend/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Localis Data API",
        "description": "This is a simple Express API based on the supplied Localis data. It currently supports retrieving all data combine and also data in the 2 original formats - length of stay data and occupancy data",
        "version": "1.0.0",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "j.hogan@qut.edu.au"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ],
...
    "paths": {
        "/api/combined_data": {
            "get": {
                "tags": [
                    "query"
                ],
                "summary": "List all data combined filtered optionally by LGAName, start and/or end",
                "description": "Full listing of all data combined from the database",
                "operationId": "listAllData",
                "parameters": [
                    {
                        "name": "LGAName",
                        "in": "query",
                        "description": "The LGA Name in question ie. Noosa, Whitsunday, Cairns, Gold Coast",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Start date filter in the format yyyy-mm-dd",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "date",
                            "default": ""
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "End date filter in the format yyyy-mm-dd",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "date",
                            "default": ""
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string",
                                    "enum": [
                                        "true",
                                        "false"
                                    ]
                                },
                                "Message": {
                                    "type": "string",
                                    "enum": [
                                        "Success",
                                        "Failure"
                                    ]
                                },
                                "combined_data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "sample_date": {
                                                "type": "date"
                                            },
                                            "lga_name": {
                                                "type": "string"
                                            },
                                            "average_historical_occupancy": {
                                                "type": "number"
                                            },
                                            "average_daily_rate": {
                                                "type": "number"
                                            },
                                            "average_length_of_stay": {
                                                "type": "number"
                                            },
                                            "average_booking_window": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/combined_data/{LGAName}": {
            "get": {
                "tags": [
                    "query"
                ],
                "summary": "List all data combined filtered by LGA Name (mandatory) and optionally start and/or end",
                "description": "Listing of combined data filtered by LGA Name from the database",
                "operationId": "listAllDataByLGA",
                "parameters": [
                    {
                        "name": "LGAName",
                        "in": "path",
                        "description": "The LGA Name in question ie. Noosa, Whitsunday, Cairns, Gold Coast",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "start",
                        "in": "query",
                        "description": "Start date filter in the format yyyy-mm-dd",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "date",
                            "default": ""
                        }
                    },
                    {
                        "name": "end",
                        "in": "query",
                        "description": "End date filter in the format yyyy-mm-dd",
                        "required": false,
                        "explode": true,
                        "schema": {
                            "type": "date",
                            "default": ""
                        }
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "Error": {
                                    "type": "string",
                                    "enum": [
                                        "true",
                                        "false"
                                    ]
                                },
                                "Message": {
                                    "type": "string",
                                    "enum": [
                                        "Success",
                                        "Failure"
                                    ]
                                },
                                "combined_data": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "properties": {
                                            "sample_date": {
                                                "type": "date"
                                            },
                                            "lga_name": {
                                                "type": "string"
                                            },
                                            "average_historical_occupancy": {
                                                "type": "number"
                                            },
                                            "average_daily_rate": {
                                                "type": "number"
                                            },
                                            "average_length_of_stay": {
                                                "type": "number"
                                            },
                                            "average_booking_window": {
                                                "type": "number"
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}